requirements.txt
boto3>=1.34.0
requests>=2.31.0
pandas>=2.1.0
pyarrow>=14.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
moto[s3,kinesis,glue,athena,dynamodb]>=5.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
